+++ src/main.rs
@@ -1,14 +1,15 @@
-mod frontend;
+mod ui;
 mod app;
 mod backend;
 
 use crate::app::Lokbuch;
 use crate::backend::database;
 
+/// Starting point of the application
 fn main() -> iced::Result {
     iced::application(Lokbuch::title, Lokbuch::update, Lokbuch::view)
         .window(iced::window::Settings {
-            icon: Some(iced::window::icon::from_file(frontend::ICON_PATH).unwrap()),
+            icon: Some(iced::window::icon::from_file(ui::ICON_PATH).unwrap()),
             ..Default::default()
         })
         .font(include_bytes!("../res/fonts/icons.ttf").as_slice())
